+++ tests/scripts/test_scan.py
@@ -2,16 +2,16 @@ import os
 os.environ['ENV'] = "test"
 import unittest
 
-from scripts.scan import get_oldest_file_in_files, get_youngest_file_in_files, scan
+from scripts.scan import get_oldest_file_in_files, get_youngest_file_in_files, scan, get_upload_options, process_upload_object
 
 # TODO: create mock files that have ages
-TESTS_PATH = os.path.expanduser("~/.onlysnarf/uploads/tests")
+TESTS_PATH = os.path.expanduser("~/.onlysnarf/uploads")
 
-class TestSelenium_Reconnect(unittest.TestCase):
+class Test_Scan(unittest.TestCase):
 
     def setUp(self):
         self.args = {
-            'action': "post",
+            'action': "test",
             'name': "",
             "oldest": False,
             "youngest": False,
@@ -20,7 +20,7 @@ class TestSelenium_Reconnect(unittest.TestCase):
             'no' : False,
             'default': False,
             'config': False,
-            'folder' : 'Tests',
+            'folder' : 'Test',
             'rootdir' : TESTS_PATH
         }
 
@@ -42,20 +42,27 @@ class TestSelenium_Reconnect(unittest.TestCase):
         self.args["config"] = True
         assert scan(self.args), "unable to scan configs"
 
-    # @unittest.skip("todo")
-    # def test_get_file_or_folder_to_upload(self):
-        # figure out how to mock each search setting: {'random': False,'oldest': False,'youngest': False,'name':False,'file':False,'folder':False}
-        # pass
+    def test_get_upload_options(self):
+        upload_config, upload_path = get_upload_options(self.args)
+        assert upload_path, "unable to get upload path via options"
 
-    @unittest.skip("todo")
+    def test_get_upload_options_config(self):
+        upload_config, upload_path = get_upload_options({**self.args, "config":True})
+        assert upload_config, "unable to get upload config"
+        assert upload_path, "unable to get upload path via config"
+
+    def test_process_upload_object(self):
+        upload_config, upload_path = get_upload_options(self.args)
+        upload_object = process_upload_object(upload_config, upload_path)
+        assert upload_object, "unable to get upload object"
+
     def test_get_oldest_file_in_files(self):
-        # use the mock directory
-        assert get_oldest_file_in_files(mock_files), "unable to get oldest file"
+        mock_files = [ '/home/skeetzo/.onlysnarf/uploads/test/post/testpost/test-image-younger.jpg', '/home/skeetzo/.onlysnarf/uploads/test/post/testpost/test-image.jpg' ]
+        self.assertEqual(get_oldest_file_in_files(mock_files), '/home/skeetzo/.onlysnarf/uploads/test/post/testpost/test-image.jpg', "unable to get oldest file")
 
-    @unittest.skip("todo")
     def test_get_youngest_file_in_files(self):
-        # use the mock directory
-        assert get_youngest_file_in_files(mock_files), "unable to get youngest file"
+        mock_files = [ '/home/skeetzo/.onlysnarf/uploads/test/post/testpost/test-image-younger.jpg', '/home/skeetzo/.onlysnarf/uploads/test/post/testpost/test-image.jpg' ]
+        self.assertEqual(get_youngest_file_in_files(mock_files), '/home/skeetzo/.onlysnarf/uploads/test/post/testpost/test-image-younger.jpg', "unable to get youngest file")
 
 ############################################################################################