+++ tests/webdriver/test_users.py
@@ -25,7 +25,7 @@ class TestWebdriver_Users(unittest.TestCase):
     @classmethod
     def tearDownClass(cls):
         configure_logs_for_module_tests(flush=True)
-        close_browser()
+        # close_browser()
 
     # def test_get_user_element(self):
         # pass